moved TestDinosaur setup to correct class
